feat: share canvas renderer for preview and export
